Add version to bigml package (bigml.__version__)
To avoid duplicating the version in various places in the code, the
version is now only stored in bigml/__init__.py. There is some hackery
in setup.py to read the version from that file without importing the
bigml package (and thus importing its dependencies, like requests, that
may not be installed yet). There's not really a good, clean way to do
this. Some options were recently discussed on Reddit:

http://www.reddit.com/r/Python/comments/tkwtt/a_comprehensive_guide_through_python_packaging/c4nljj6