Changing checks on create functions
The create functions first check that the origin resources are
finished a limited amount of times. If they aren't still ready
we let them be used as origin resources in the create call,
that will fail with the corresponding message.

We also update the resource_is_ready function to accept HTTP_ACCEPTED as
a correct code for finished resources.